fix(secrets): macOS errors during build
